requirements.txt
fastapi
uvicorn
sqlmodel
pytz
pyyaml
pytest
httpx
python-telegram-bot
sqlalchemy
pydantic